test(reconciliation): keep engine checks within source budget
