+++ src/tavily_proxy/proxy_quota_sync_and_jobs.rs
@@ -3010,66 +3010,3 @@ fn normalize_quota_sync_fetch_error(err: ProxyError) -> ProxyError {
         other => other,
     }
 }
-
-#[cfg(test)]
-mod reconciliation_engine_tests {
-    use std::sync::atomic::AtomicI64;
-
-    use super::{
-        ReconciliationEngine, ReconciliationOutcome, should_emit_reconciliation_summary_at,
-    };
-
-    #[test]
-    fn reconciliation_summary_logging_is_limited_to_one_per_minute() {
-        let last_emitted_at = AtomicI64::new(0);
-
-        assert!(should_emit_reconciliation_summary_at(&last_emitted_at, 1_000));
-        assert!(!should_emit_reconciliation_summary_at(&last_emitted_at, 1_059));
-        assert!(should_emit_reconciliation_summary_at(&last_emitted_at, 1_060));
-    }
-
-    #[test]
-    fn non_success_outcomes_do_not_clear_upstream_429_state() {
-        assert!(!ReconciliationEngine::clears_upstream_429(
-            ReconciliationOutcome::TransportFailure
-        ));
-        assert!(!ReconciliationEngine::clears_upstream_429(
-            ReconciliationOutcome::SemanticFailure
-        ));
-        assert!(!ReconciliationEngine::clears_upstream_429(
-            ReconciliationOutcome::LocalPressure
-        ));
-        assert!(ReconciliationEngine::clears_upstream_429(
-            ReconciliationOutcome::Settled
-        ));
-    }
-
-    #[test]
-    fn successful_terminal_outcomes_clear_both_backoffs() {
-        assert!(ReconciliationEngine::clears_local_pressure(
-            ReconciliationOutcome::Settled
-        ));
-        assert!(ReconciliationEngine::clears_local_pressure(
-            ReconciliationOutcome::NoAdjustment
-        ));
-        assert!(ReconciliationEngine::clears_upstream_429(
-            ReconciliationOutcome::NoAdjustment
-        ));
-    }
-
-    #[test]
-    fn failure_outcomes_prevent_a_same_round_success_from_clearing_429() {
-        assert_eq!(
-            ReconciliationEngine::outcome(1, 1, 0, false, false, true, false),
-            Some(ReconciliationOutcome::TransportFailure)
-        );
-        assert_eq!(
-            ReconciliationEngine::outcome(1, 0, 0, false, false, false, true),
-            Some(ReconciliationOutcome::SemanticFailure)
-        );
-        assert_eq!(
-            ReconciliationEngine::outcome(1, 1, 0, false, true, false, false),
-            Some(ReconciliationOutcome::LocalPressure)
-        );
-    }
-}
